#70 - separating the Auth/auth event config beans from the listeners. This is separating the jobs and making sure only Beans are in @Configuration classes.
